+++ services/services/player/client/player_client.h
@@ -55,7 +55,7 @@ public:
     int32_t GetCurrentTime(int32_t &currentTime) override;
     int32_t GetPlaybackPosition(int32_t &playbackPosition) override;
     int32_t GetCurrentPresentationTimestamp(int64_t &currentPresentation) override;
-    int32_t Seek(int32_t mSeconds, PlayerSeekMode mode) override;
+    int32_t Seek(int64_t mSeconds, PlayerSeekMode mode) override;
     int32_t SeekToDefaultPosition() override;
     int32_t GetSubtitleTrackInfo(std::vector<Format> &subtitleTrack) override;
     int32_t GetPlaybackSpeed(PlaybackRateMode &mode) override;
