+++ test/fuzztest/player_fuzztest/playerstub_fuzzer/player_service_proxy_fuzzer.h
@@ -103,7 +103,7 @@ public:
     {
         return 0;
     }
-    int32_t Seek(int32_t mSeconds, PlayerSeekMode mode) override
+    int32_t Seek(int64_t mSeconds, PlayerSeekMode mode) override
     {
         return 0;
     }
