+++ test/unittest/timeout_test/mock/mock_i_player_service.h
@@ -47,7 +47,7 @@ public:
     MOCK_METHOD(int32_t, ReleaseSync, (), (override));
     MOCK_METHOD(int32_t, SetVolume, (float leftVolume, float rightVolume), (override));
     MOCK_METHOD(int32_t, SetVolumeMode, (int32_t mode), (override));
-    MOCK_METHOD(int32_t, Seek, (int32_t mSeconds, PlayerSeekMode mode), (override));
+    MOCK_METHOD(int32_t, Seek, (int64_t mSeconds, PlayerSeekMode mode), (override));
     MOCK_METHOD(int32_t, GetCurrentTime, (int32_t &currentTime), (override));
     MOCK_METHOD(int32_t, GetPlaybackPosition, (int32_t &playbackPosition), (override));
     MOCK_METHOD(int32_t, GetVideoTrackInfo, (std::vector<Format> &videoTrack), (override));
